requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
apscheduler==3.10.4
itsdangerous==2.2.0
starlette==0.37.2
pydantic==2.7.1
pydantic[email]==2.7.1
python-dotenv==1.0.1
pillow==10.3.0
aiofiles==23.2.1
